+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 ### Breaking
 * Stop ignoring whitespace. I . . . actually don't remember why I did this. It's not necessary for numbers, and it just makes it impossible to store strings with prefix/suffix whitespace, which breaks the guaranteed roundtrip requirements. If you're reading this in to-be-released notes then let me know if it breaks your workflow, otherwise this is probably going live.
 
+### Improved
+* Error message with multiple nodes of a single name.
+
 ### Fixed
 * String roundtripping with leading/trailing whitespace.
 * Exceptions thrown in Recorder could propagate outside Dec.